sentiment/requirements.txt
streamlit
pandas
groq
pdfplumber
python-dotenv
matplotlib